feature: add compilations business logic

+++ main-service/src/main/java/ru/practicum/mainservice/model/request/UpdateCompilationRequest.java
@@ -11,7 +11,7 @@ import java.util.List;
 @AllArgsConstructor
 @NoArgsConstructor
 public class UpdateCompilationRequest {
-    List<Long> events;
+    List<Integer> events;
     Boolean pinned;
     String title;
